requirements.txt
aiofiles==23.2.1
altgraph==0.17.4
async-tkinter-loop==0.9.3
auto-py-to-exe==2.42.0
AwesomeTkinter==2021.11.8
Babel==2.14.0
beautifulsoup4==4.12.3
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
CTkMessagebox==2.5
CTkSpinbox==1.3.0
customtkinter==5.2.2
darkdetect==0.8.0
dbus-python==1.3.2
distlib==0.3.8
Eel==0.16.0
future==0.18.3
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==3.0.3
idna==3.6
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.4
packaging==23.2
patchelf==0.17.2.1
pbkdf2==1.3
pillow==10.2.0
plyer==2.1.0
psutil==5.9.8
pygame==2.5.2
pyglet==2.0.10
Pygments==2.17.2
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.1
pyparsing==3.1.1
pyperclip==1.8.2
python-bidi==0.4.2
pytube==15.0.0
requests==2.31.0
requests_download==0.1.2
rich==13.7.0
shellingham==1.5.4
six==1.16.0
soupsieve==2.5
tkfilebrowser==2.3.2
typer==0.9.0
typing_extensions==4.9.0
urllib3==2.2.0
websockets==12.0
whichcraft==0.6.1
wifi==0.3.8
yarg==0.1.9
zope.event==5.0
zope.interface==6.1
